enh: restrict api_key authentication
now restricted to webservices rules and git smart.
to enable it for your controller, add a "sub begin" with
$c->stash->{api_key_authentication} = 1 in it (for the whole controller
file in this case).
also added a new api_key_authentication config option for enabling a
global api_key login function, not recommended though, useful for
panicking situation in legacy installations