enh: include_into fieldlet is now in the palette
before it was a required field, and it was broken due to the fact that
it did not have a default section (ie 'details') so it would not show. The model
function it calls was broken too, it was using the category fieldlets data.

So now the whole thing is a palette fieldlet, and it's up to the user if
he wants to add it to a form rule.